rename module for consistency
